requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
openai==1.12.0
httpx==0.26.0
python-dotenv==1.0.0
openpyxl==3.1.2
nest-asyncio==1.5.8
tiktoken==0.7.0
pydantic==2.5.0
SQLAlchemy==2.0.29
